perf: dynamic import ShareBadgePreview, memo BadgePreviewCard, wrap lottie experiment. Fixes #125. Fixes #126. Fixes #131
